Refactoring (split test into seperate classes)
