+++ src/test/java/chapterSix/Login.java
@@ -1,0 +1,19 @@
+package chapterSix;
+
+import general.TestShopScenario;
+import org.testng.annotations.Test;
+
+/**
+ * Created by Daniel on 21/07/2017.
+ */
+public class Login extends TestShopScenario{
+    @Test
+    public void logInSuccesFull(){
+        goToHomePage();
+        maximizeWindow();
+        goToLoginPage();
+        loginFlow();
+        assertSuccesfullLogin();
+        goToHomePage();
+    }
+}
